Move all cross-language transform registrar code to one file (read CR)
