--- sdks/java/io/kinesis/src/main/java/org/apache/beam/sdk/io/kinesis/CrossLanguageConfiguration.java
@@ -1,46 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.beam.sdk.io.kinesis;
-
-public abstract class CrossLanguageConfiguration {
-  String streamName;
-  String awsAccessKey;
-  String awsSecretKey;
-  String region;
-  String serviceEndpoint;
-
-  public void setStreamName(String streamName) {
-    this.streamName = streamName;
-  }
-
-  public void setAwsAccessKey(String awsAccessKey) {
-    this.awsAccessKey = awsAccessKey;
-  }
-
-  public void setAwsSecretKey(String awsSecretKey) {
-    this.awsSecretKey = awsSecretKey;
-  }
-
-  public void setRegion(String region) {
-    this.region = region;
-  }
-
-  public void setServiceEndpoint(String serviceEndpoint) {
-    this.serviceEndpoint = serviceEndpoint;
-  }
-}
